Separate requests schema and add first test
