+++ src/lib/airtable/tables/requests.js
@@ -1,6 +1,6 @@
 const { merge } = require("lodash");
 const _ = require("lodash");
-const { airbase } = require("~airtable/bases");
+const { table } = require("./requestsSchema")
 
 const requestNotInSlack = (r) => {
   const meta = r.get(fields.meta);
@@ -22,7 +22,7 @@ const notForDrivingCluster = (r) => {
 exports.deleteRequest = async (recordId) => {
   console.log("Deleting record");
   try {
-    const records = await requestsTable.destroy([recordId]);
+    const records = await table.destroy([recordId]);
     return [records[0], null];
   } catch (e) {
     console.error(`Couldn't delete request: ${e}`);
@@ -34,7 +34,7 @@ exports.createRequest = async (request) => {
   // TODO: add asserts for non-|| fields below
   console.log("creating record");
   try {
-    const record = await requestsTable.create({
+    const record = await table.create({
       [fields.message]: request.message,
       [fields.phone]: request.phone || "",
       [fields.type]: request.source,
@@ -53,7 +53,7 @@ exports.createRequest = async (request) => {
 
 exports.findRequestByExternalId = async (externalId) => {
   try {
-    const record = await requestsTable
+    const record = await table
       .select({
         filterByFormula: `({${fields.externalId}} = '${externalId}')`,
       })
@@ -78,7 +78,7 @@ exports.findDeliveryNeededRequests = async () => {
   );
   const formula = `OR(${statusConstraints.join(", ")})`;
   try {
-    const requests = await requestsTable
+    const requests = await table
       .select({
         filterByFormula: formula,
       })
@@ -101,7 +101,7 @@ exports.findOpenRequestsForSlack = async () => {
   );
   const formula = `OR(${statusConstraints.join(", ")})`;
   try {
-    const requests = await requestsTable
+    const requests = await table
       .select({
         filterByFormula: formula,
       })
@@ -121,7 +121,7 @@ exports.findRequestByCode = async (code) => {
     return [null, `Request code must be at least 4 characters.`];
   }
   try {
-    const records = await requestsTable
+    const records = await table
       .select({
         filterByFormula: `(FIND('${code}', {${fields.code}}) > 0)`,
       })
@@ -138,7 +138,7 @@ exports.findRequestByCode = async (code) => {
 
 exports.findRequestByPhone = async (phone) => {
   try {
-    const records = await requestsTable
+    const records = await table
       .select({
         maxRecords: 1,
         fields: [fields.phone],
@@ -165,7 +165,7 @@ exports.updateRequestByCode = async (code, update) => {
     return [null, `Request code must be at least 4 characters.`];
   }
   try {
-    const records = await requestsTable
+    const records = await table
       .select({
         filterByFormula: `(FIND('${code}', {${fields.code}}) > 0)`,
       })
@@ -185,7 +185,7 @@ exports.updateRequestByCode = async (code, update) => {
       id: record.id,
       fields: update,
     };
-    const updatedRecords = await requestsTable.update([airUpdate]);
+    const updatedRecords = await table.update([airUpdate]);
     return [updatedRecords[0], null];
   } catch (e) {
     return [null, `Error while processing update: ${e}`];
@@ -196,7 +196,7 @@ exports.unlinkSlackMessage = async (slackTs, slackChannel) => {
   const tsFilter = `"slack_ts":"${slackTs}"`;
   const channelFilter = `"slack_channel":"${slackChannel}"`;
   const filter = `AND(SEARCH('${tsFilter}', {${fields.meta}}), SEARCH('${channelFilter}', {${fields.meta}}))`;
-  await requestsTable
+  await table
     .select({
       filterByFormula: filter,
     })
@@ -205,7 +205,7 @@ exports.unlinkSlackMessage = async (slackTs, slackChannel) => {
         const meta = removeSlackMeta(record.get(fields.meta));
 
         try {
-          await requestsTable.update([
+          await table.update([
             { id: record.id, fields: { [fields.meta]: meta } },
           ]);
         } catch (e) {
@@ -224,108 +224,3 @@ const removeSlackMeta = (meta) =>
     .omit("slack_channel")
     .thru(JSON.stringify)
     .value();
-
-// ==================================================================
-// Schema
-// ==================================================================
-
-const requestsTableName = (exports.tableName = "Requests");
-const requestsTable = (exports.table = airbase(requestsTableName));
-const fields = (exports.fields = {
-  phone: "Phone",
-  time: "Time",
-  type: "Text or Voice?",
-  type_options: {
-    text: "text",
-    voice: "voice",
-    manyc: "manyc",
-    email: "email",
-  },
-  message: "Message",
-  crossStreetFirst: "Cross Street #1",
-  crossStreetSecond: "Cross Street #2",
-  email: "Email Address",
-  neighborhoodArea: "Neighborhood Area (See Map)",
-  neighborhoodArea_options: {
-    ne: "NE",
-    nw: "NW",
-    se: "SE",
-    sw: "SW",
-    notCrownHeights: "Other - not Crown Heights",
-  },
-  languages: "Languages",
-  languages_options: {
-    spanish: "Spanish",
-    chineseMandarin: "Chinese - Mandarin",
-    chineseCantonese: "Chinese - Cantonese",
-    haitianKreyol: "Haitian Kreyol",
-    russian: "Russian",
-    bengali: "Bengali",
-    french: "French",
-    yiddish: "Yiddish",
-    italian: "Italian",
-    korean: "Korean",
-    arabic: "Arabic",
-    tagalog: "Tagalog",
-    polish: "Polish",
-    asl: "ASL",
-    english: "English",
-    eglish: "eglish",
-  },
-  supportType: "What type(s) of support are you seeking?",
-  supportType_options: {
-    delivery: "Deliver groceries or supplies to me",
-    prescriptionPickUp: "Pick up a prescription for me",
-    "1On1CheckInsPhoneCallZoomEtcToTouchBaseWithANeighbor":
-      "1 on 1 check-ins (phone call, Zoom, etc to touch base with a neighbor)",
-    financialSupport: "Financial support",
-    translation:
-      "Translation and interpretation into a language other than English",
-    socialServices:
-      "Social Services guidance (filing for medicare, unemployment, etc)",
-    other: "Other (please describe below)",
-  },
-  financialSupportNeeded: "Financial Support Needed?",
-  financialSupportNeeded_options: {
-    yes: "yes - need donation",
-    no: "no donation need",
-  },
-  intakeVolunteer: "Intake volunteer",
-  deliveryVolunteer: "Delivery volunteer",
-  timeSensitivity: "Time Sensitivity",
-  intakeNotes: "Intake General Notes",
-  startEditingHere: "Start Editing Here!",
-  receipts: "Receipts",
-  code: "Code",
-  status: "Status",
-  status_options: {
-    dispatchNeeded: "Dispatch Needed",
-    dispatchStarted: "Dispatch Started",
-    noAnswer: "No Answer (call back)",
-    deliveryNeeded: "Delivery Needed",
-    duplicate: "Duplicate",
-    deliveryAssigned: "Delivery Assigned",
-    requestComplete: "Request Complete",
-    beyondCrownHeights: "Beyond Crown Heights",
-    otherFollowup: "Other Followup",
-    sentToAnotherGroup: "Sent to Another Group",
-    needsPosting: "Needs Posting!",
-    brownsville: "Brownsville/East NY",
-  },
-  externalId: "External Id",
-  deliverySlackId: "Delivery slackId",
-  lastModified: "Last Modified",
-  meta: "Meta",
-  lastProcessed: "Last Processed",
-  firstName: "First Name",
-  triggerBackfill: "Trigger Backfill",
-  neighborhood: "Neighborhood MA-NYC",
-  householdSize: "Household Size",
-  forDrivingClusters: "For Driving Clusters",
-});
-exports.SENSITIVE_FIELDS = [
-  fields.phone,
-  fields.email,
-  fields.message,
-  fields.intakeNotes,
-];
